Mengintegrasikan menu pendaftaran dan alternatif

+++ app/Http/Controllers/Menu/PendaftaranController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\User;
 use App\Models\Menu\DataKegiatan;
 use App\Models\Menu\Pendaftaran;
+use App\Models\Menu\DataAlternatif;
 use App\Http\Controllers\Controller; 
 use App\Http\Requests\DataPendaftaranRequest;  
 use Illuminate\Support\Facades\Auth;
@@ -26,11 +27,22 @@ class PendaftaranController extends Controller
         return view('menu.pendaftaran.create', compact('users', 'kegiatans','pendaftarans'));
     }
 
-    public function store(DataPendaftaranRequest $request)
+    public function store(Request $request)
     {
-        $data = $request->validated();
-        $data['id_data_user'] = Auth::user()->id;
-        Pendaftaran::create($data);
+        $validatedData = $request->validate([
+            'id_data_user' => 'required|exists:users,id',
+            'id_data_kegiatan' => 'required|exists:data_kegiatan,id',
+            'provinsi' => 'required',
+            'kabupaten_kota' => 'required',
+            'jabatan' => 'required'
+        ]);
+
+        $pendaftaran = Pendaftaran::create($validatedData);
+
+        $data_alternatif = DataAlternatif::create([
+            'id_pendaftaran' => $pendaftaran->id
+        ]);
+
         return redirect()->route('pendaftaran')->with('success', 'Data pendaftaran berhasil dibuat');
     }
 
@@ -48,18 +60,41 @@ class PendaftaranController extends Controller
         return view('menu.pendaftaran.edit', compact('pendaftaran', 'users', 'kegiatans'));
     }
 
-    public function update(DataPendaftaranRequest $request, $id)
+    public function update(Request $request, Pendaftaran $pendaftaran)
     {
-        $pendaftaran = Pendaftaran::findOrFail($id);
-        $pendaftaran->update($request->all());
+        $validatedData = $request->validate([
+            'id_data_user' => 'required|exists:users,id',
+            'id_data_kegiatan' => 'required|exists:data_kegiatan,id',
+            'provinsi' => 'required',
+            'kabupaten_kota' => 'required',
+            'jabatan' => 'required'
+        ]);
+
+        $pendaftaran->update($validatedData);
+
+        $data_alternatif = $pendaftaran->data_alternatif;
+        $data_alternatif->update([
+            'id_pendaftaran' => $pendaftaran->id
+        ]);
+
         return redirect()->route('pendaftaran')->with('success', 'Data pendaftaran berhasil diupdate');
     }
 
-    public function destroy($id)
-    {
-        $pendaftaran = Pendaftaran::findOrFail($id);
+    public function destroy(Pendaftaran $pendaftaran)
+    {  
+    $data_alternatif = $pendaftaran->data_alternatif;
+
+    if ($data_alternatif) {
+        $data_alternatif->delete();
+    }
+
+    if ($pendaftaran) {
         $pendaftaran->delete();
         return redirect()->route('pendaftaran')->with('success', 'Data pendaftaran berhasil dihapus');
+    } else {
+        return redirect()->route('pendaftaran')->with('error', 'Data pendaftaran tidak ditemukan');
     }
+}
+
 
 }
